refactor(prisma): move the admin seed out of core infrastructure
seedAdmin creates users and assigns ADMIN/USER roles, which is business
knowledge, and it lived under core/prisma/ next to the database plumbing.
No code in src/ ever imported it: its only consumers are the two root seed
scripts, so prisma/ is both its nearest common ancestor and the conventional
home for Prisma seed tooling.

Extend the vitest unit project to prisma/**/*.spec.ts as part of the move.
The colocated spec left src/ with the file it covers, and without this the
suite would have silently dropped from 82 tests to 78.
